+++ ts/features/idpay/initiative/details/components/InitiativeSettingsComponent.tsx
@@ -1,7 +1,7 @@
 import { useNavigation } from "@react-navigation/core";
 import * as O from "fp-ts/lib/Option";
 import { pipe } from "fp-ts/lib/function";
-import { List, ListItem } from "native-base";
+import { ListItem as NBListItem } from "native-base";
 import React from "react";
 import { View } from "react-native";
 import Placeholder from "rn-placeholder";
@@ -22,6 +22,7 @@ import {
   IOStackNavigationProp
 } from "../../../../../navigation/params/AppParamsList";
 import customVariables from "../../../../../theme/variables";
+import { IDPayUnsubscriptionRoutes } from "../../../unsubscription/navigation/navigator";
 import { IDPayConfigurationRoutes } from "../../configuration/navigation/navigator";
 
 type Props = {
@@ -33,42 +34,42 @@ type SettingsButtonProps = {
   subtitle?: string;
   onPress?: () => void;
   hasWarnings?: boolean;
+  isLoading?: boolean;
 };
 
 const SettingsButtonComponent = (props: SettingsButtonProps) => {
-  const { title, subtitle, onPress, hasWarnings } = props;
+  const { title, subtitle, onPress, hasWarnings, isLoading } = props;
 
-  const subtitleComponent = pipe(
-    subtitle,
-    O.fromNullable,
-    O.fold(
-      () => (
+  const getSubtitleComponent = () => {
+    if (isLoading) {
+      return (
         <>
           <VSpacer size={4} />
           <Placeholder.Box animate="fade" height={16} width={120} radius={4} />
         </>
-      ),
-      subtitle => {
-        if (hasWarnings) {
-          <LabelSmall weight="SemiBold" color="red">
-            {I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.configured.settings.actionsRequired"
-            )}
-          </LabelSmall>;
-        }
+      );
+    }
 
-        return (
-          <LabelSmall weight="Regular" color="bluegrey">
-            {subtitle}
-          </LabelSmall>
-        );
-      }
-    )
-  );
+    if (hasWarnings) {
+      return (
+        <LabelSmall weight="SemiBold" color="red">
+          {I18n.t(
+            "idpay.initiative.details.initiativeDetailsScreen.configured.settings.actionsRequired"
+          )}
+        </LabelSmall>
+      );
+    }
+
+    return (
+      <LabelSmall weight="Regular" color="bluegrey">
+        {subtitle}
+      </LabelSmall>
+    );
+  };
 
   return (
-    <ListItem onPress={onPress} style={{ paddingEnd: 0 }}>
-      {props.hasWarnings && (
+    <NBListItem onPress={onPress} style={{ paddingEnd: 0 }}>
+      {hasWarnings && (
         <>
           <IconFont name={"io-warning"} color={IOColors.red} />
           <HSpacer size={16} />
@@ -76,17 +77,17 @@ const SettingsButtonComponent = (props: SettingsButtonProps) => {
       )}
       <View style={IOStyles.flex}>
         <H4>{title}</H4>
-        {subtitleComponent}
+        {getSubtitleComponent()}
       </View>
       <IconFont
         name={"io-right"}
         color={customVariables.contentPrimaryBackground}
       />
-    </ListItem>
+    </NBListItem>
   );
 };
 
-export const InitiativeSettingsComponent = (props: Props) => {
+const InitiativeSettingsComponent = (props: Props) => {
   const { initiative } = props;
 
   const navigation = useNavigation<IOStackNavigationProp<AppParamsList>>();
@@ -110,6 +111,16 @@ export const InitiativeSettingsComponent = (props: Props) => {
     });
   };
 
+  const navigateToUnsubscription = (
+    initiativeId: string,
+    initiativeName?: string
+  ) => {
+    navigation.navigate(IDPayUnsubscriptionRoutes.IDPAY_UNSUBSCRIPTION_MAIN, {
+      initiativeId,
+      initiativeName
+    });
+  };
+
   const instrumentsSettingsButton = pipe(
     initiative,
     O.fromNullable,
@@ -119,6 +130,7 @@ export const InitiativeSettingsComponent = (props: Props) => {
           title={I18n.t(
             "idpay.initiative.details.initiativeDetailsScreen.configured.settings.associatedPaymentMethods"
           )}
+          isLoading={true}
         />
       ),
       ({ initiativeId, nInstr, status }) => (
@@ -152,6 +164,7 @@ export const InitiativeSettingsComponent = (props: Props) => {
           title={I18n.t(
             "idpay.initiative.details.initiativeDetailsScreen.configured.settings.selectedIBAN"
           )}
+          isLoading={true}
         />
       ),
       ({ initiativeId, iban, status }) => (
@@ -169,6 +182,21 @@ export const InitiativeSettingsComponent = (props: Props) => {
     )
   );
 
+  const unsubscriptionButton = pipe(
+    initiative,
+    O.fromNullable,
+    O.fold(
+      () => null,
+      ({ initiativeId, initiativeName }) => (
+        <SettingsButtonComponent
+          title={"Rimuovi iniziativa"}
+          onPress={() => navigateToUnsubscription(initiativeId, initiativeName)}
+          subtitle=""
+        />
+      )
+    )
+  );
+
   return (
     <>
       <H3>
@@ -177,10 +205,12 @@ export const InitiativeSettingsComponent = (props: Props) => {
         )}
       </H3>
       <VSpacer size={8} />
-      <List>
-        {instrumentsSettingsButton}
-        {ibanSettingsButton}
-      </List>
+      {instrumentsSettingsButton}
+      {ibanSettingsButton}
+      {/* TODO: temporary button, removed in IODPAY-175  */}
+      {unsubscriptionButton}
     </>
   );
 };
+
+export { InitiativeSettingsComponent };