+++ CHANGELOG.md
@@ -2,6 +2,71 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.31.0-rc.3](https://github.com/pagopa/io-app/compare/2.31.0-rc.2...2.31.0-rc.3) (2023-05-04)
+
+
+### Bug Fixes
+
+* [[IABT-1473](https://pagopa.atlassian.net/browse/IABT-1473)] Fix contextual help when user is not authenticated ([#4585](https://github.com/pagopa/io-app/issues/4585)) ([c094044](https://github.com/pagopa/io-app/commit/c0940441a1605f26574ef6507c2c158b4370cf43))
+
+
+### Chores
+
+* [[IODPAY-216](https://pagopa.atlassian.net/browse/IODPAY-216)] Add onboarding PDND criteria values ([#4558](https://github.com/pagopa/io-app/issues/4558)) ([0758907](https://github.com/pagopa/io-app/commit/0758907efc58f5d9c3d0eaa0289d18ec25f08e87))
+
+## [2.31.0-rc.2](https://github.com/pagopa/io-app/compare/2.31.0-rc.1...2.31.0-rc.2) (2023-05-03)
+
+
+### Features
+
+* **Firma con IO:** [[SFEQS-1378](https://pagopa.atlassian.net/browse/SFEQS-1378)] Update main app navigator stack to enable FCI deeplink ([#4568](https://github.com/pagopa/io-app/issues/4568)) ([0326230](https://github.com/pagopa/io-app/commit/03262307c64b04320d665ae3eaf0aaa194b1a688))
+* **Firma con IO:** [[SFEQS-1576](https://pagopa.atlassian.net/browse/SFEQS-1576)] Update router screen to render error based on the response status code ([#4566](https://github.com/pagopa/io-app/issues/4566)) ([da3cf0b](https://github.com/pagopa/io-app/commit/da3cf0ba74610669615c235f504c329aa4d5fc7a)), closes [/github.com/pagopa/io-dev-api-server/blob/e5363f1a9afb60ccb0488b5bb708437d23eb8682/src/routers/features/fci/index.ts#L29](https://github.com/pagopa//github.com/pagopa/io-dev-api-server/blob/e5363f1a9afb60ccb0488b5bb708437d23eb8682/src/routers/features/fci/index.ts/issues/L29)
+* **Firma con IO:** [[SFEQS-800](https://pagopa.atlassian.net/browse/SFEQS-800)] Add mixpanel FCI events ([#4570](https://github.com/pagopa/io-app/issues/4570)) ([8e1a539](https://github.com/pagopa/io-app/commit/8e1a5395bebc4a78df0369fc21a6ceea205bab1a))
+* [[IOCOM-172](https://pagopa.atlassian.net/browse/IOCOM-172),[IOCOM-173](https://pagopa.atlassian.net/browse/IOCOM-173),[IOCOM-179](https://pagopa.atlassian.net/browse/IOCOM-179)] Display remote precondition before opening the message details ([#4565](https://github.com/pagopa/io-app/issues/4565)) ([86b9d20](https://github.com/pagopa/io-app/commit/86b9d2057c3ce6be625b5348d9e2d24712a1e991)), closes [/github.com/pagopa/io-dev-api-server/blob/IOCOM-171-remote-bottom-sheet-message/src/config.ts#L131](https://github.com/pagopa//github.com/pagopa/io-dev-api-server/blob/IOCOM-171-remote-bottom-sheet-message/src/config.ts/issues/L131) [/github.com/pagopa/io-dev-api-server/blob/IOCOM-171-remote-bottom-sheet-message/src/config.ts#L78](https://github.com/pagopa//github.com/pagopa/io-dev-api-server/blob/IOCOM-171-remote-bottom-sheet-message/src/config.ts/issues/L78)
+* **IDPay:** [[IODPAY-171](https://pagopa.atlassian.net/browse/IODPAY-171)] IDPay initiative unsubscription ([#4462](https://github.com/pagopa/io-app/issues/4462)) ([c3c2f67](https://github.com/pagopa/io-app/commit/c3c2f675ebeb0efca9a3c9a9ddfb0ab1e1f555ef))
+
+
+### Bug Fixes
+
+* **Firma con IO:** [[SFEQS-1635](https://pagopa.atlassian.net/browse/SFEQS-1635)] Fix mixpanel events action ([#4580](https://github.com/pagopa/io-app/issues/4580)) ([e1e534c](https://github.com/pagopa/io-app/commit/e1e534c665c11df9dbac2d26f6e88d69f9ccc577))
+* **IDPay:** [[IODPAY-149](https://pagopa.atlassian.net/browse/IODPAY-149),[IODPAY-138](https://pagopa.atlassian.net/browse/IODPAY-138),[IODPAY-139](https://pagopa.atlassian.net/browse/IODPAY-139),[IODPAY-142](https://pagopa.atlassian.net/browse/IODPAY-142),[IODPAY-184](https://pagopa.atlassian.net/browse/IODPAY-184)] Show operation details bottom sheet from timeline ([#4370](https://github.com/pagopa/io-app/issues/4370)) ([d6dc1de](https://github.com/pagopa/io-app/commit/d6dc1def8b84a101772ff76b511c5d0169dfc4ff))
+* [[IABT-1471](https://pagopa.atlassian.net/browse/IABT-1471)] Update ZenDesk SDKs to support iOS 16 ([#4577](https://github.com/pagopa/io-app/issues/4577)) ([d67f147](https://github.com/pagopa/io-app/commit/d67f1471717cdbde05482c38970bd8008fba94c5))
+
+
+### Chores
+
+* **IDPay:** [[IODPAY-196](https://pagopa.atlassian.net/browse/IODPAY-196)] Update new initiative placeholder screen ([#4578](https://github.com/pagopa/io-app/issues/4578)) ([b3f1259](https://github.com/pagopa/io-app/commit/b3f1259027906717c1e74ef4354fa9710d324f53))
+* [[IOAPPFD0-81](https://pagopa.atlassian.net/browse/IOAPPFD0-81)] Update README.md ([#4528](https://github.com/pagopa/io-app/issues/4528)) ([4262807](https://github.com/pagopa/io-app/commit/426280728f887928246a90a8c13739ec390946fe))
+* **IDPay:** [[IODPAY-195](https://pagopa.atlassian.net/browse/IODPAY-195)] New IDPay bonus card in wallet ([#4554](https://github.com/pagopa/io-app/issues/4554)) ([080aecc](https://github.com/pagopa/io-app/commit/080aecce3cc564e7db1e20328823e04621daed2f))
+* **IDPay:** [[IODPAY-223](https://pagopa.atlassian.net/browse/IODPAY-223)] Initiative details screen refactoring ([#4581](https://github.com/pagopa/io-app/issues/4581)) ([46eb50a](https://github.com/pagopa/io-app/commit/46eb50a289aa9ddfa86d6a060da3f0043c19062f))
+* [[IOAPPFD0-64](https://pagopa.atlassian.net/browse/IOAPPFD0-64)] Enable deep-link handling in react-navigation ([#4152](https://github.com/pagopa/io-app/issues/4152)) ([563bafe](https://github.com/pagopa/io-app/commit/563bafec9141cb87607e950c5c242949701df1ce))
+* **IDPay:** [[IODPAY-221](https://pagopa.atlassian.net/browse/IODPAY-221)] Update IDPay API definition to v2.52.1 ([#4575](https://github.com/pagopa/io-app/issues/4575)) ([0a4cc8d](https://github.com/pagopa/io-app/commit/0a4cc8d7b05165cd5091a9340ffa00d368956245))
+
+## [2.31.0-rc.1](https://github.com/pagopa/io-app/compare/2.31.0-rc.0...2.31.0-rc.1) (2023-04-27)
+
+
+### Features
+
+* [[IOCOM-164](https://pagopa.atlassian.net/browse/IOCOM-164)] Replaces custom types and decoders on io-backend client ([#4546](https://github.com/pagopa/io-app/issues/4546)) ([49d0f9b](https://github.com/pagopa/io-app/commit/49d0f9b833986b64c9e23f19ef4f1a3f76f67c1b))
+* **IDPay:** [[IODPAY-116](https://pagopa.atlassian.net/browse/IODPAY-116)] Add info modal in IBAN landing screen ([#4556](https://github.com/pagopa/io-app/issues/4556)) ([4719766](https://github.com/pagopa/io-app/commit/4719766b6aa5dbde640e875cf37c8c93e7efb204))
+
+
+### Bug Fixes
+
+* **IDPay:** [[IODPAY-192](https://pagopa.atlassian.net/browse/IODPAY-192)] Wrong copy in initiativeDetails' "lastUpdated" ([#4569](https://github.com/pagopa/io-app/issues/4569)) ([3ce719b](https://github.com/pagopa/io-app/commit/3ce719b4cd0296537860309c21a07eae698ea414))
+* [[IOAPPFD0-91](https://pagopa.atlassian.net/browse/IOAPPFD0-91)] Message sender details not wrapping ([#4567](https://github.com/pagopa/io-app/issues/4567)) ([fe664f6](https://github.com/pagopa/io-app/commit/fe664f6f2abcbce558353b9f6640dd9df0436037))
+* **IDPay:** [[IODPAY-212](https://pagopa.atlassian.net/browse/IODPAY-212)] "Tap to scroll" floating button in IDPay onboarding screen ([#4541](https://github.com/pagopa/io-app/issues/4541)) ([da01632](https://github.com/pagopa/io-app/commit/da0163209d8fca1b98fd19c3ca34f763bc2f1441))
+* **IDPay:** [[IODPAY-217](https://pagopa.atlassian.net/browse/IODPAY-217)] Fix IDPay initiative settings component skeleton ([#4564](https://github.com/pagopa/io-app/issues/4564)) ([13fbb19](https://github.com/pagopa/io-app/commit/13fbb19b5817de680d873ed206a784fbbb54e7e2))
+* [[IOAPPFD0-89](https://pagopa.atlassian.net/browse/IOAPPFD0-89)] Update slack channel of CIE iOS button check ([#4562](https://github.com/pagopa/io-app/issues/4562)) ([3a2ea1d](https://github.com/pagopa/io-app/commit/3a2ea1dad45b980ba35187e807178e3f9e3b008d))
+
+
+### Chores
+
+* [[IOAPPFD0-80](https://pagopa.atlassian.net/browse/IOAPPFD0-80)] Upgrades react-native to 0.69.9 ([#4514](https://github.com/pagopa/io-app/issues/4514)) ([75c44c4](https://github.com/pagopa/io-app/commit/75c44c44ed7b38c4fa320fa6d21c22e64d9e4f6a))
+* [[IOPID-167](https://pagopa.atlassian.net/browse/IOPID-167)] TestAuthenticationScreen UI Refinement ([#4571](https://github.com/pagopa/io-app/issues/4571)) ([db35afe](https://github.com/pagopa/io-app/commit/db35afe7f31dd9ed098d5c75ec0d119e84dfd7c8)), closes [/github.com/pagopa/io-app/pull/4571/files#diff-b215994f68a368f1dc322f668b2937aeede527cbaa9242f10a9896526d1982](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-b215994f68a368f1dc322f668b2937aeede527cbaa9242f10a9896526d1982) [/github.com/pagopa/io-app/pull/4571/files#diff-7d599e5ee6dea19a2231e4a0e43ca1cc92d8d9680a2c6d7390e44999da2baf10](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-7d599e5ee6dea19a2231e4a0e43ca1cc92d8d9680a2c6d7390e44999da2baf10) [/github.com/pagopa/io-app/pull/4571/files#diff-54190256bc3cfec139b2b7e5e70c221f05d09fa8c9650ec95a3b38f6c9c36d91](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-54190256bc3cfec139b2b7e5e70c221f05d09fa8c9650ec95a3b38f6c9c36d91) [/github.com/pagopa/io-app/pull/4571/files#diff-8f675a5e21b9a99d7ec5d3e2bb2cf950abb463a4e0e9525b5286035567ca87a8](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-8f675a5e21b9a99d7ec5d3e2bb2cf950abb463a4e0e9525b5286035567ca87a8) [/github.com/pagopa/io-app/pull/4571/files#diff-6286d90afa4e0d1ffb5531596558295d753f7b1d0831d4d1e8448eef9d90fd32](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-6286d90afa4e0d1ffb5531596558295d753f7b1d0831d4d1e8448eef9d90fd32) [/github.com/pagopa/io-app/pull/4571/files#diff-67816cf28cc48675238d98a2f9eb645014f1c8c4dd5cea88c7cb6478056a274](https://github.com/pagopa//github.com/pagopa/io-app/pull/4571/files/issues/diff-67816cf28cc48675238d98a2f9eb645014f1c8c4dd5cea88c7cb6478056a274)
+* **Firma con IO:** [[SFEQS-1629](https://pagopa.atlassian.net/browse/SFEQS-1629)] Add zendesk fci custom field id ([#4572](https://github.com/pagopa/io-app/issues/4572)) ([6994fb2](https://github.com/pagopa/io-app/commit/6994fb27828850a5a3236ca3b3cf30806002180b))
+* **IDPay:** [[IODPAY-214](https://pagopa.atlassian.net/browse/IODPAY-214)] IDPay Initiative status labels in details screen ([#4557](https://github.com/pagopa/io-app/issues/4557)) ([ab0b594](https://github.com/pagopa/io-app/commit/ab0b594bb46e2ecac039fc9695726cd7ed4b7872))
+
 ## [2.31.0-rc.0](https://github.com/pagopa/io-app/compare/2.30.0-rc.2...2.31.0-rc.0) (2023-04-21)